requirements.txt
scrapy==1.4.0
configparser==3.5.0
python-telegram-bot==5.0.0